Update README.md
How It Works
The inaugural piece in a groundbreaking NFT art collection meticulously crafted to immortalize the diverse cultures thriving within the heart of the Amazon Rainforest. This mesmerizing digital artwork encapsulates the spirit, traditions, and resilience of the Rainforest's peoples, serving as a poignant testament to their harmonious coexistence with nature.

What sets "Rainforest Unveiled" apart is its profound impact in the territory. A remarkable 80% of the proceeds generated from this NFT will be directed towards a local rural community nestled in the Amazon, empowering them to preserve their heritage and sustainably manage their environment.

As a token of gratitude for your support, the acquisition of "Rainforest Unveiled" also grants you an exclusive membership to the G Club. This membership unlocks a multitude of benefits, offering you privileged access to the Rainforest Online Boutique, GREEN-C. Immerse yourself in ethically sourced, eco-friendly products, allowing you to make a tangible contribution to the preservation of the Amazon while enjoying the finest offerings from this enchanting realm.

By embracing "Roots Unveiled" and becoming a G Club member, you not only gain ownership of a unique digital masterpiece but actively participate in the conservation of the Amazon Rainforest and its extraordinary communities.

MVP link: https://www.green-c.io/get-benefits